Move safe onboarding to new folder
