+++ src/steps/safe-onboarding/styled.tsx
